addRule finished, creation class Sequence
