+++ src/com/practiceProject/LSystem.java
@@ -28,8 +28,9 @@ public class LSystem extends AbstractLSystem{
     public void addRule(Symbol sym, String expansion) {
         Sequence sequence = new Sequence();
         sequence.getString(expansion);
-        List<Symbol.Seq> listSequence = new LinkedList();
-
+        List<Symbol.Seq> symbolRule = new LinkedList<>();
+        symbolRule = sequence.getList();
+        rules.put(sym, symbolRule);
     }
 
     @Override
@@ -65,43 +66,5 @@ public class LSystem extends AbstractLSystem{
     @Override
     public Rectangle2D tell(Turtle turtle, Symbol sym, int rounds) {
         return null;
-    }
-
-    public class Sequence implements Symbol.Seq {
-
-        String expansion;
-
-        @Override
-        public Iterator<Symbol> iterator() {
-            return new ListIterator<Symbol>(this);
-        }
-
-        @Override
-        public void add(Symbol data) {
-
-        }
-
-        @Override
-        public void getString(String s) {
-            expansion = s;
-        }
-
-        @Override
-        public Symbol getHead() {
-            if(expansion == null){
-                return null;
-            }
-            Symbol s = alphabet.get(expansion.charAt(0));
-            return s;
-        }
-
-        @Override
-        public Symbol getTail() {
-            if(expansion == null){
-                return null;
-            }
-            Symbol s = alphabet.get(expansion.charAt(expansion.length()-1));
-            return s;
-        }
     }
 }
